nlp-to-strategy-engine/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# LLM and parsing
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0

# Validation and data models
pydantic>=2.0.0

# DSL parsing
lark>=1.1.0

# Backtesting
ta>=0.11.0  # Technical analysis library

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
